+++ saleor/core/permissions.py
@@ -3,6 +3,7 @@ from django.contrib.auth.models import Permission
 MODELS_PERMISSIONS = [
     "account.manage_users",
     "account.manage_staff",
+    "account.manage_service_accounts",
     "account.impersonate_users",
     "discount.manage_discounts",
     "giftcard.manage_gift_card",