+++ saleor/graphql/payment/schema.py
@@ -1,7 +1,7 @@
 import graphene
-from graphql_jwt.decorators import permission_required
 
 from ..core.fields import PrefetchingConnectionField
+from ..decorators import permission_required
 from .enums import PaymentGatewayEnum
 from .mutations import PaymentCapture, PaymentRefund, PaymentSecureConfirm, PaymentVoid
 from .resolvers import resolve_payment_client_token, resolve_payments