+++ tests/api/conftest.py
@@ -8,7 +8,7 @@ from django.shortcuts import reverse
 from django.test.client import MULTIPART_CONTENT, Client
 from graphql_jwt.shortcuts import get_token
 
-from saleor.account.models import User
+from saleor.account.models import ServiceAccount, User
 
 from .utils import assert_no_permission
 
@@ -19,16 +19,22 @@ class ApiClient(Client):
     """GraphQL API client."""
 
     def __init__(self, *args, **kwargs):
-        user = kwargs.pop("user")
+        user = kwargs.pop("user", AnonymousUser())
+        service_account = kwargs.pop("service_account", None)
         self.user = user
+        self.service_token = None
         if not user.is_anonymous:
             self.token = get_token(user)
+        elif service_account:
+            self.service_token = service_account.auth_token
         super().__init__(*args, **kwargs)
 
     def _base_environ(self, **request):
         environ = super()._base_environ(**request)
         if not self.user.is_anonymous:
             environ.update({"HTTP_AUTHORIZATION": "JWT %s" % self.token})
+        elif self.service_token:
+            environ.update({"HTTP_AUTHORIZATION": "Bearer %s" % self.service_token})
         return environ
 
     def post(self, data=None, **kwargs):
@@ -86,6 +92,11 @@ class ApiClient(Client):
 
 
 @pytest.fixture
+def service_account_api_client(service_account):
+    return ApiClient(service_account=service_account)
+
+
+@pytest.fixture
 def staff_api_client(staff_user):
     return ApiClient(user=staff_user)
 
@@ -130,3 +141,8 @@ def user_list_not_active(user_list):
     users = User.objects.filter(pk__in=[user.pk for user in user_list])
     users.update(is_active=False)
     return users
+
+
+@pytest.fixture
+def service_account(db):
+    return ServiceAccount.objects.create(name="Sample service account", is_active=True)